+++ panel/__init__.py
@@ -24,7 +24,7 @@ from . import widgets # noqa
 from .config import config, panel_extension as extension, __version__ # noqa
 from .depends import bind, depends # noqa
 from .interact import interact # noqa
-from .io import _jupyter_server_extension_paths, ipywidget, serve, state # noqa
+from .io import _jupyter_server_extension_paths, ipywidget, state # noqa
 from .layout import ( # noqa
     Accordion, Card, Column, GridSpec, GridBox, FlexBox, Tabs, Row,
     Spacer, WidgetBox
