+++ panel/viewable.py
@@ -23,6 +23,7 @@ from bokeh.io import curdoc as _curdoc
 from pyviz_comms import JupyterCommManager
 
 from .config import config, panel_extension
+from .io.document import init_doc
 from .io.embed import embed_state
 from .io.loading import start_loading_spinner, stop_loading_spinner
 from .io.model import add_to_doc, patch_cds_msg
@@ -31,7 +32,6 @@ from .io.notebook import (
 )
 from .io.save import save
 from .io.state import state
-from .io.server import init_doc, serve
 from .util import escape, param_reprs
 
 
@@ -378,6 +378,7 @@ class ServableMixin(object):
           Returns the Bokeh server instance or the thread the server
           was launched on (if threaded=True)
         """
+        from .io.server import serve
         return serve(
             self, port=port, address=address, websocket_origin=websocket_origin,
             show=open, start=True, title=title, verbose=verbose,
